server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
chromadb
pypdf
python-dotenv
